requirements.txt
# Autotok - Videos Requirements

# Google Sheets integration
gspread==6.0.0
oauth2client==4.1.3

# Anthropic Claude API
anthropic==0.25.0

# Google Cloud Text-to-Speech
google-cloud-texttospeech==2.16.3

# Creatomate API
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pandas==2.2.0

# Logging
colorlog==6.8.2
